different modes for points info calculation

+++ calcPointsInfo/pointsInfoStruct.hpp
@@ -6,7 +6,8 @@
 struct PointsInfo {
     // that's an array and not a matrix, because it will be easier to work with array when we will apply intrinsics
     float*     lastPointRadiusMatrix;
-    size_t*    escTimesMatrix;
+    int*       escTimesMatrix; // won't need size_t, int is more than enough, also it's size is 8, which is
+                               // very convenient as we use intrinsics
 };
 
 #endif
